- Moved map location related files to its own package
