+++ src/main/use_case/navigation/GraphPathFinder.java
@@ -11,6 +11,7 @@ import org.jgrapht.graph.SimpleWeightedGraph;
 import data_access.LocationDataAccess;
 import entity.Location;
 import entity.Room;
+import use_case.navigation.MapLocation.MapLocation;
 
 /**
  * A class that finds the shortest path between two locations on the map.
