+++ src/main/use_case/navigation/MapLocation/ImageMapLocationFactory.java
@@ -1,0 +1,11 @@
+package use_case.navigation.MapLocation;
+
+/**
+ * Factory for creating ImageMapLocation objects.
+ */
+public class ImageMapLocationFactory implements MapLocationFactory {
+    @Override
+    public MapLocation createMapLocation(String id, double mapX, double mapY, int floor) {
+        return new ImageMapLocation(id, mapX, mapY, floor);
+    }
+}
