+++ src/main/use_case/navigation/NavigationOutputData.java
@@ -1,5 +1,7 @@
 package use_case.navigation;
 
+import use_case.navigation.MapLocation.MapLocation;
+
 import java.util.ArrayList;
 import java.util.List;
